refactor: move function for initializing monitoring off to the side.
