+++ src/thumbnailer/util.py
@@ -1,8 +1,16 @@
-from logging import getLogger, basicConfig, DEBUG, INFO, debug, info
+from os import environ
+from logging import getLogger, basicConfig, DEBUG, INFO, debug, info, warning
+
+from sentry_sdk import init
+from sentry_sdk.integrations.aws_lambda import AwsLambdaIntegration
+
+from thumbnailer import version
 
 
 DEFAULT_WIDTH = 222
 DEFAULT_HEIGHT = 222
+MONITORING_DSN = 'SENTRY_DSN'
+OPERATING_ENV = 'OPERATING_ENV'
 
 
 def configure_logging(verbose):
@@ -22,3 +30,24 @@ def configure_logging(verbose):
             datefmt='%Y/%m/%d %H:%M:%S',
             level=level,
         )
+
+
+def init_monitoring():
+    dsn = environ.get(MONITORING_DSN)
+    env = environ.get(OPERATING_ENV)
+    
+    if not dsn:
+        warning(f'DSN not found in envronment under key {MONITORING_DSN}')
+        return
+    
+    info(f'Configuring monitoring via DSN: {dsn}')
+    init(
+        dsn=dsn,
+        integrations=[AwsLambdaIntegration()],
+        release=f'v{version}',
+        send_default_pii=False,
+        traces_sample_rate=0.50,
+        environment=env,
+        _experiments={'auto_enabling_integrations': True},
+    )
+
